+++ yandex/resource_yandex_storage_bucket.go
@@ -1,6 +1,7 @@
 package yandex
 
 import (
+	"bytes"
 	"fmt"
 	"log"
 	"reflect"
@@ -12,8 +13,10 @@ import (
 	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/aws/aws-sdk-go/service/s3"
 	"github.com/hashicorp/go-getter/helper/url"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/hashcode"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/validation"
 )
 
 func resourceYandexStorageBucket() *schema.Resource {
@@ -60,9 +63,51 @@ func resourceYandexStorageBucket() *schema.Resource {
 			},
 
 			"acl": {
-				Type:     schema.TypeString,
-				Default:  "private",
-				Optional: true,
+				Type:          schema.TypeString,
+				Default:       "private",
+				Optional:      true,
+				ConflictsWith: []string{"grant"},
+			},
+
+			"grant": {
+				Type:          schema.TypeSet,
+				Optional:      true,
+				Set:           grantHash,
+				ConflictsWith: []string{"acl"},
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"id": {
+							Type:     schema.TypeString,
+							Optional: true,
+						},
+						"type": {
+							Type:     schema.TypeString,
+							Required: true,
+							ValidateFunc: validation.StringInSlice([]string{
+								s3.TypeCanonicalUser,
+								s3.TypeGroup,
+							}, false),
+						},
+						"uri": {
+							Type:     schema.TypeString,
+							Optional: true,
+						},
+
+						"permissions": {
+							Type:     schema.TypeSet,
+							Required: true,
+							Set:      schema.HashString,
+							Elem: &schema.Schema{
+								Type: schema.TypeString,
+								ValidateFunc: validation.StringInSlice([]string{
+									s3.PermissionFullControl,
+									s3.PermissionRead,
+									s3.PermissionWrite,
+								}, false),
+							},
+						},
+					},
+				},
 			},
 
 			"cors_rule": {
@@ -212,6 +257,12 @@ func resourceYandexStorageBucketUpdate(d *schema.ResourceData, meta interface{})
 		}
 	}
 
+	if d.HasChange("grant") {
+		if err := resourceYandexStorageBucketGrantsUpdate(s3Client, d); err != nil {
+			return err
+		}
+	}
+
 	return resourceYandexStorageBucketRead(d, meta)
 }
 
@@ -332,6 +383,28 @@ func resourceYandexStorageBucketRead(d *schema.ResourceData, meta interface{}) e
 			return err
 		}
 	}
+
+	//Read the Grant ACL. Reset if `acl` (canned ACL) is set.
+	if acl, ok := d.GetOk("acl"); ok && acl.(string) != "private" {
+		if err := d.Set("grant", nil); err != nil {
+			return fmt.Errorf("error resetting grant %s", err)
+		}
+	} else {
+		apResponse, err := retryFlakyS3Responses(func() (interface{}, error) {
+			return s3Client.GetBucketAcl(&s3.GetBucketAclInput{
+				Bucket: aws.String(d.Id()),
+			})
+		})
+		if err != nil {
+			return fmt.Errorf("error getting getting storage (%s) ACL: %s", d.Id(), err)
+		}
+		log.Printf("[DEBUG] getting storage: %s, read ACL grants policy: %+v", d.Id(), apResponse)
+		grants := flattenGrants(apResponse.(*s3.GetBucketAclOutput))
+		if err := d.Set("grant", schema.NewSet(grantHash, grants)); err != nil {
+			return fmt.Errorf("error setting grant %s", err)
+		}
+	}
+
 
 	return nil
 }
@@ -497,7 +570,7 @@ func resourceYandexStorageBucketCORSUpdate(s3Client *s3.S3, d *schema.ResourceDa
 			err = waitCorsPut(s3Client, bucket, corsConfiguration)
 		}
 		if err != nil {
-			return fmt.Errorf("error putting storage CORS: %s", err)
+			return fmt.Errorf("error putting bucket CORS: %s", err)
 		}
 	}
 
@@ -825,6 +898,171 @@ func validateS3BucketName(value string) error {
 	}
 	if !regexp.MustCompile(`^[0-9a-zA-Z-.]+$`).MatchString(value) {
 		return fmt.Errorf("only alphanumeric characters, hyphens, and periods allowed in %q", value)
+	}
+
+	return nil
+}
+
+func grantHash(v interface{}) int {
+	var buf bytes.Buffer
+	m, ok := v.(map[string]interface{})
+
+	if !ok {
+		return 0
+	}
+
+	if v, ok := m["id"]; ok {
+		buf.WriteString(fmt.Sprintf("%s-", v.(string)))
+	}
+	if v, ok := m["type"]; ok {
+		buf.WriteString(fmt.Sprintf("%s-", v.(string)))
+	}
+	if v, ok := m["uri"]; ok {
+		buf.WriteString(fmt.Sprintf("%s-", v.(string)))
+	}
+	if p, ok := m["permissions"]; ok {
+		buf.WriteString(fmt.Sprintf("%v-", p.(*schema.Set).List()))
+	}
+	return hashcode.String(buf.String())
+}
+
+func resourceYandexStorageBucketGrantsUpdate(s3conn *s3.S3, d *schema.ResourceData) error {
+	bucket := d.Get("bucket").(string)
+	rawGrants := d.Get("grant").(*schema.Set).List()
+
+	if len(rawGrants) == 0 {
+		log.Printf("[DEBUG] Storage bucket: %s, Grants fallback to canned ACL", bucket)
+		if err := resourceYandexStorageBucketACLUpdate(s3conn, d); err != nil {
+			return fmt.Errorf("Error fallback to canned ACL, %s", err)
+		}
+	} else {
+		apResponse, err := retryFlakyS3Responses(func() (interface{}, error) {
+			return s3conn.GetBucketAcl(&s3.GetBucketAclInput{
+				Bucket: aws.String(d.Id()),
+			})
+		})
+
+		if err != nil {
+			return fmt.Errorf("error getting Storage Bucket (%s) ACL: %s", d.Id(), err)
+		}
+
+		ap := apResponse.(*s3.GetBucketAclOutput)
+		log.Printf("[DEBUG] Storage bucket: %s, read ACL grants policy: %+v", d.Id(), ap)
+
+		grants := make([]*s3.Grant, 0, len(rawGrants))
+		for _, rawGrant := range rawGrants {
+			log.Printf("[DEBUG] Storage bucket: %s, put grant: %#v", bucket, rawGrant)
+			grantMap := rawGrant.(map[string]interface{})
+			permissions := grantMap["permissions"].(*schema.Set).List()
+			if err := validateBucketPermissions(permissions); err != nil {
+				return err
+			}
+			for _, rawPermission := range permissions {
+				ge := &s3.Grantee{}
+				if i, ok := grantMap["id"].(string); ok && i != "" {
+					ge.SetID(i)
+				}
+				if t, ok := grantMap["type"].(string); ok && t != "" {
+					ge.SetType(t)
+				}
+				if u, ok := grantMap["uri"].(string); ok && u != "" {
+					ge.SetURI(u)
+				}
+
+				g := &s3.Grant{
+					Grantee:    ge,
+					Permission: aws.String(rawPermission.(string)),
+				}
+				grants = append(grants, g)
+			}
+		}
+
+		grantsInput := &s3.PutBucketAclInput{
+			Bucket: aws.String(bucket),
+			AccessControlPolicy: &s3.AccessControlPolicy{
+				Grants: grants,
+				Owner:  ap.Owner,
+			},
+		}
+
+		log.Printf("[DEBUG] Bucket: %s, put Grants: %#v", bucket, grantsInput)
+
+		_, err = retryFlakyS3Responses(func() (interface{}, error) {
+			return s3conn.PutBucketAcl(grantsInput)
+		})
+
+		if err != nil {
+			return fmt.Errorf("Error putting bucket Grants: %s", err)
+		}
+	}
+	return nil
+}
+
+func flattenGrants(ap *s3.GetBucketAclOutput) []interface{} {
+	//if ACL grants contains bucket owner FULL_CONTROL only - it is default "private" acl
+	if len(ap.Grants) == 1 && aws.StringValue(ap.Grants[0].Grantee.ID) == aws.StringValue(ap.Owner.ID) &&
+		aws.StringValue(ap.Grants[0].Permission) == s3.PermissionFullControl {
+		return nil
+	}
+
+	getGrant := func(grants []interface{}, grantee map[string]interface{}) (interface{}, bool) {
+		for _, pg := range grants {
+			pgt := pg.(map[string]interface{})
+			if pgt["type"] == grantee["type"] && pgt["id"] == grantee["id"] && pgt["uri"] == grantee["uri"] &&
+				pgt["permissions"].(*schema.Set).Len() > 0 {
+				return pg, true
+			}
+		}
+		return nil, false
+	}
+
+	grants := make([]interface{}, 0, len(ap.Grants))
+	for _, granteeObject := range ap.Grants {
+		grantee := make(map[string]interface{})
+		grantee["type"] = aws.StringValue(granteeObject.Grantee.Type)
+
+		if granteeObject.Grantee.ID != nil {
+			grantee["id"] = aws.StringValue(granteeObject.Grantee.ID)
+		}
+		if granteeObject.Grantee.URI != nil {
+			grantee["uri"] = aws.StringValue(granteeObject.Grantee.URI)
+		}
+		if pg, ok := getGrant(grants, grantee); ok {
+			pg.(map[string]interface{})["permissions"].(*schema.Set).Add(aws.StringValue(granteeObject.Permission))
+		} else {
+			grantee["permissions"] = schema.NewSet(schema.HashString, []interface{}{aws.StringValue(granteeObject.Permission)})
+			grants = append(grants, grantee)
+		}
+	}
+
+	return grants
+}
+
+func validateBucketPermissions(permissions []interface{}) error {
+	var (
+		fullControl bool
+		permissionRead bool
+		permissionWrite bool
+	)
+
+	for _, p := range permissions {
+		s := p.(string)
+		switch s {
+		case s3.PermissionFullControl:
+			fullControl = true
+		case s3.PermissionRead:
+			permissionRead = true
+		case s3.PermissionWrite:
+			permissionWrite = true
+		}
+	}
+
+	if fullControl && len(permissions) > 1 {
+		return fmt.Errorf(`should not use other ACP permissions along with "FULL_CONTROL" permission`)
+	}
+
+	if permissionWrite && !permissionRead {
+		return fmt.Errorf(`Please always provide "READ" permission, when granting "WRITE"`)
 	}
 
 	return nil