+++ CHANGELOG.md
@@ -9,6 +9,7 @@ ENHANCEMENTS:
 * managed-kubernetes: add `node_ipv4_cidr_mask_size` attribute to `yandex_kubernetes_cluster` resource and data source
 * managed-kubernetes: add `deploy_policy.max_unavailable` and `deploy_policy.max_expansion` attributes to `yandex_kubernetes_node_group` resource and data source
 * serverless: add `environment` attribute in `yandex_function` resource and data source
+* storage: support custom acl grants for `yandex_storage_bucket` 
 
 BUG FIXES:
 * mdb: fix modifying `yandex_mdb_mysql_cluster` attribute `backup_window_start`